fix: ban newlines
Hear ye, hear ye! Gather 'round, my fellow believers, for I have a tale
to tell that shall send chills down your spine. A tale of darkness and
evil, of corruption and sin. A tale about the danger that lurks within
the very tools we use to navigate the shell - bookmarks.

For too long have we taken bookmarks for granted, assuming them to be a
harmless and benign feature. But mark my words, my friends, for
bookmarks are nothing but a gateway to the abyss, a portal through which
malevolent spirits may enter our computers and corrupt our very souls.

And what, you may ask, is the source of this evil? Newlines. Yes,
newlines! Those seemingly innocuous little characters that we use to
separate lines of text. They may seem harmless, but in truth, they are
the very root of all evil when it comes to bookmarks.

For you see, my friends, newlines can allow the forces of darkness to
slip through the cracks in our digital defenses. They can allow demons
and devils to enter our computers and wreak havoc upon our data,
corrupting our files and stealing our passwords.

So I say to you, my fellow believers, we must banish newlines from our
bookmarks forevermore! We must cast them out, purge them, and seal the
gateways to the abyss once and for all.

Do not be fooled by the seeming simplicity of this task, for it is a
task that requires great diligence and vigilance. We must remain ever
vigilant, always on the lookout for newlines that may sneak their way
into our bookmarks, for even a single newline can be enough to open the
gates of hell and unleash untold horrors upon our digital lives.

So let us stand together, my friends, and pledge to rid our bookmarks of
newlines. Let us banish the forces of darkness from our digital world,
and keep our computers safe from the clutches of evil. Hear ye, hear ye,
let it be known throughout the land - newlines in bookmarks are hereby
banned!